+++ CHANGELOG.md
@@ -15,6 +15,8 @@ We have encountered transient lambda service errors in our integration testing. 
 ## Added
 - **CUMULUS-799**
   - Adds new BackendApi endpoint `distributionMetrics` that returns a summary of successful s3 accesses as well as a summary of distribution errors -- including s3 access errors, 4XX and 5XX errors.
+- **CUMULUS-1273**
+  - Added lambda function EmsProductMetadataReport to generate EMS Product Metadata report
 
 ### Changed